+++ bench/gar-riccati.cpp
@@ -21,7 +21,7 @@ const uint nu = 12;
 static void BM_serial(benchmark::State &state) {
   uint horz = (uint)state.range(0);
   VectorXs x0 = VectorXs::NullaryExpr(nx, normal_unary_op{});
-  const LQRProblemTpl<double> problem = generate_problem(x0, horz, nx, nu);
+  const LqrProblemTpl<double> problem = generate_problem(x0, horz, nx, nu);
   ProximalRiccatiSolver<double> solver(problem);
   const double mu = 1e-11;
   auto [xs, us, vs, lbdas] = lqrInitializeSolution(problem);
@@ -35,7 +35,7 @@ static void BM_serial(benchmark::State &state) {
 template <uint NPROC> static void BM_parallel(benchmark::State &state) {
   uint horz = (uint)state.range(0);
   VectorXs x0 = VectorXs::NullaryExpr(nx, normal_unary_op{});
-  LQRProblemTpl<double> problem = generate_problem(x0, horz, nx, nu);
+  LqrProblemTpl<double> problem = generate_problem(x0, horz, nx, nu);
   ParallelRiccatiSolver<double> solver(problem, NPROC);
   const double mu = 1e-11;
   auto [xs, us, vs, lbdas] = lqrInitializeSolution(problem);
@@ -49,7 +49,7 @@ template <uint NPROC> static void BM_parallel(benchmark::State &state) {
 static void BM_stagedense(benchmark::State &state) {
   uint horz = (uint)state.range(0);
   VectorXs x0 = VectorXs::NullaryExpr(nx, normal_unary_op{});
-  LQRProblemTpl<double> problem = generate_problem(x0, horz, nx, nu);
+  LqrProblemTpl<double> problem = generate_problem(x0, horz, nx, nu);
   RiccatiSolverDense<double> solver(problem);
   const double mu = 1e-11;
   auto [xs, us, vs, lbdas] = lqrInitializeSolution(problem);